% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_input_crime_raw}
\alias{example_input_crime_raw}
\title{Example input crime data}
\format{
A data frame of 1 row and 9 columns:
\describe{
\item{\code{case_id}}{A unique reference number for the input crime. Can be a character or numeric vector.}
\item{\code{offence_type}}{The type of input crime. A factor with levels "burglary", "robbery" or "sex", being the crimes GP-SMART is calibrated for use with.}
\item{\code{x}}{Easting coordinate. Must be in metres to enable distance calculations.}
\item{\code{y}}{Northing coordinate. Must be in metres to enable distance calculations.}
\item{\code{start_date}}{Start date of the crime. A date in format "YYYY-MM-DD".}
\item{\code{end_date}}{End date of the crime, if the exact date is not known. A date in format "YYYY-MM-DD".}
\item{\code{start_time}}{Start time of the crime. A difftime in format "HH:MM:SS".}
\item{\code{end_time}}{End time of the crime, if the exact time is not known. A difftime in format "HH:MM:SS".}
\item{\code{location_type}}{The type of location in which the crime was committed. A factor with levels "residential", "commercial", "public", "street" or "unknown".}
}
}
\source{
Based on crime data provided by New Zealand Police
}
\usage{
data(example_input_crime_raw)
}
\description{
A dataset containing an example input crime (a burglary), in raw form.
It includes the minimum information needed for the function \code{\link[=fn_prepare_input_crime]{fn_prepare_input_crime()}} to operate.
This example is fictional; the location, time, case id, and location type are made up,
but are typical of the crimes in the data used in Curtis-Ham et al (2022) to calibrate and test GP-SMART.
}
\references{
Curtis-Ham S., Bernasco, W., Medvedev, O. N., & Polaschek, D. L. L (2022).
'A new geographic profiling method for mapping and ranking suspects in crime investigations: GP-SMART'.
Journal of Investigative Psychology and Offender Profiling. https://doi.org/10.1002/jip.1585
}
\keyword{datasets}
