% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_input_suspects_raw}
\alias{example_input_suspects_raw}
\title{Example input suspect data}
\format{
A data frame of 14826 row and 14 columns:
\describe{
\item{\code{person_id}}{A unique reference number for the suspect. Can be a character or numeric vector.}
\item{\code{death_date}}{Suspect's date of death if deceased. A date in format "YYYY-MM-DD"}
\item{\code{x}}{Easting coordinate. Must be in metres to enable distance calculations.}
\item{\code{y}}{Northing coordinate. Must be in metres to enable distance calculations.}
\item{\code{node_category}}{The general category of the activity location. A character vector with values "span" or "event". Span and event nodes are treated differently when calculating activity node attributes.}
\item{\code{node_type}}{The type of activity location. A factor with levels "home", "family_immediate", "family_ip", "family_other", "school", "work", "offence", "victim_witness", "incident", "police_other". Node type determines some activity node attribute values.}
\item{\code{prior_offence_type}}{The type of offence, if it's an offence activity location and one of the crime types for which GP-SMART is currently calibrated. A factor with levels "burglary", "robbery" or "sex", being the crimes GP-SMART is calibrated for. Used for calculating behaviour similarity.}
\item{\code{span_start_date}}{Start date of the activity location, if a span node. A date in format "YYYY-MM-DD".}
\item{\code{span_end_date}}{End date of the activity location, if an event node. A date in format "YYYY-MM-DD".}
\item{\code{event_start_date}}{Start date of the event, if an event node. A date in format "YYYY-MM-DD".}
\item{\code{event_end_date}}{End date of the event, if an event node. A date in format "YYYY-MM-DD".}
\item{\code{event_start_time}}{Start time of the event, if an event node. A difftime in format "HH:MM:SS". Time is not present for "police_other" event nodes.}
\item{\code{event_end_time}}{End time of the event, if an event node. A difftime in format "HH:MM:SS". Time is not present for "police_other" event nodes.}
\item{\code{location_type}}{The type of location in which the crime was committed. A factor with levels "residential", "commercial", "public", "street" or "unknown".}
}
}
\source{
Based on Offender activity location data provided by New Zealand Police.
}
\usage{
data(example_input_suspects_raw)
}
\description{
A dataset containing an example set of suspect activity locations (nodes), in raw form.
It includes the minimum information needed for the function \code{\link[=fn_prepare_suspect_data]{fn_prepare_suspect_data()}} to operate.
The example data are fictional; they were simulated based on the suspect activity location data used in Curtis-Ham et al (2022) to calibrate and test GP-SMART.
The example data reflect the general location and distribution of the original data (in space, in time, per node type and per suspect).
They do not represent the exact locations and times of suspects' activity locations in the original data.
The simulation process meant that node and location types do not necessarily match the map location of the nodes.
For example, a node with location type "residential" might fall on a commercial premises.
In the original data, nodes with location type "residential" fall on residential premises.
}
\references{
Curtis-Ham S., Bernasco, W., Medvedev, O. N., & Polaschek, D. L. L (2022).
'A new geographic profiling method for mapping and ranking suspects in crime investigations: GP-SMART'.
Journal of Investigative Psychology and Offender Profiling.
}
\keyword{datasets}
