% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_map_gpsmart_output.R
\name{fn_map_gpsmart_output}
\alias{fn_map_gpsmart_output}
\title{Map the output of GP-SMART for an input crime and shortlisted suspects}
\usage{
fn_map_gpsmart_output(gpsmart_output, coordinate_system = 2193)
}
\arguments{
\item{gpsmart_output}{A list with four elements, returned by running the function \code{\link[=fn_gpsmart]{fn_gpsmart()}} with \code{return_node_predictions} set to \code{TRUE}:
\describe{
\item{\code{suspects_ranked}}{A data frame with n rows = n suspects with nodes within the \code{search_radius} of the input crime,
ranked by their maximum predicted crime probability.}
\item{\code{node_predictions}}{A data frame with n rows = n suspect nodes within \code{search_radius},
including all node attributes and predicted crime probability per node.}
\item{\code{input_crime}}{A data frame with 1 row and 16 columns created by \code{\link[=fn_prepare_input_crime]{fn_prepare_input_crime()}},
entered as input for \code{\link[=fn_gpsmart]{fn_gpsmart()}}.}
\item{\code{search_radius}}{The search radius (in kilometres), entered as input for \code{\link[=fn_gpsmart]{fn_gpsmart()}}.}
}}

\item{coordinate_system}{CRS number for the coordinate system for the X and Y coordinates of the input crime and suspect node data.
The default is New Zealand Transverse Mercator (2193).}
}
\value{
An interactive map object that can be opened in the RStudio Viewer or an internet browser and saved as an html file for sharing.
The map displays:
\itemize{
\item{the location of the input crime (IC);}
\item{the activity nodes of all shortlisted suspects with nodes within the \code{search_radius} when running \code{\link[=fn_gpsmart]{fn_gpsmart()}}, coloured by their predicted probability (as a percentile);}
\item{the search radius.}
}
The user can:
\itemize{
\item{pan and zoom to specific locations;}
\item{search (enter text or use drop-down menu) to filter nodes to a particular suspect(s) based on their identifier;}
\item{use a slider to limit the displayed nodes to those of the top x ranked suspects;}
\item{use tick boxes to limit the displayed nodes according to \code{node_type} and other node attributes used in \code{\link[=fn_gpsmart]{fn_gpsmart()}} to calculate the predicted probability;}
\item{hover over a node to show the rank of the suspect;}
\item{click on a node to show a popup box with the suspect's rank and identifier, the \code{node_type}, and the node's values on the GP-SMART variables;}
\item{search for an address or place name.}
}
}
\description{
Creates an interactive map showing shortlisted suspects' activity locations (nodes) within the user's set search radius of an input crime.
Use \code{\link[htmltools:save_html]{htmltools::save_html()}} to save the output map as an html file.
}
\details{
\itemize{
\item{The function will return an error if there are no node predictions in \code{gpsmart_output} to map.}
\item{Requires an internet connection.}
}
}
\examples{
\dontrun{
data(example_input_crime_raw)
data(example_input_suspects_raw)
gpsmart_output <- fn_gpsmart(
  input_crime = fn_prepare_input_crime(example_input_crime_raw),
  input_suspects = fn_prepare_suspect_data(example_input_suspects_raw),
  search_radius = 10,
  return_node_predictions = TRUE
)
fn_map_gpsmart_output(
  gpsmart_output,
  coordinate_system = 2193
)
}

}
\seealso{
\itemize{
\item \code{\link[=fn_gpsmart]{fn_gpsmart()}} creates the input object for this function.
}
}
\author{
Sophie Curtis-Ham
}
