% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prepare_input_crime.R
\name{fn_prepare_input_crime}
\alias{fn_prepare_input_crime}
\title{Prepare input crime for GP-SMART}
\usage{
fn_prepare_input_crime(input_crime_raw)
}
\arguments{
\item{input_crime_raw}{A data frame with 1 row and 9 columns:
\describe{
\item{\code{case_id}}{A unique reference number for the input crime. Can be a character or numeric vector. Used to name gp-smart outputs.}
\item{\code{offence_type}}{The type of input crime. A factor with levels "burglary", "robbery" or "sex", being the crimes GP-SMART is calibrated for use with.}
\item{\code{x}}{Easting coordinate. Must be in metres to enable distance calculations.}
\item{\code{y}}{Northing coordinate. Must be in metres to enable distance calculations.}
\item{\code{start_date}}{Start date of the crime. A date in format "YYYY-MM-DD".}
\item{\code{end_date}}{End date of the crime, if the exact date is not known. A date in format "YYYY-MM-DD".}
\item{\code{start_time}}{Start time of the crime. A difftime in format "HH:MM:SS".}
\item{\code{end_time}}{End time of the crime, if the exact time is not known. A difftime in format "HH:MM:SS".}
\item{\code{location_type}}{The type of location in which the crime was committed. A factor with levels "residential", "commercial", "public", "street" or "unknown".}
}}
}
\value{
A data frame the same as the input with 7 additional columns:
\itemize{
\item{\code{offence_subtype_ic}}{The subtype of input crime, abbreviated (resburg, nonresburg, comrob, persrob, sex).}
\item{\code{daypart_ic}}{The day part of the input crime (daytime, evening, night).}
\item{\code{weekpart_ic}}{The week part of the input crime (Weekday, weekend).}
\item{\code{spring_ic}}{Whether the input crime occurred in spring (1 = yes, 0 = no).}
\item{\code{summer_ic}}{Whether the input crime occurred in summer (1 = yes, 0 = no).}
\item{\code{autumn_ic}}{Whether the input crime occurred in autumn (1 = yes, 0 = no).}
\item{\code{winter_ic}}{Whether the input crime occurred in winter (1 = yes, 0 = no).}
}
}
\description{
Checks that the minimum necessary variables are present and calculates additional attributes needed for \code{\link[=fn_gpsmart]{fn_gpsmart()}}.
}
\details{
\itemize{
\item{The function will return an error if the input does not contain the correct columns in the correct format.}
\item{Season variables are based on Southern Hemisphere.}
}
}
\examples{
data(example_input_crime_raw)
fn_prepare_input_crime(example_input_crime_raw)

}
\seealso{
\itemize{
\item \code{\link[=fn_prepare_suspect_data]{fn_prepare_suspect_data()}} checks that the minimum necessary variables are present and creates the \code{input_suspects} data frame for use in \code{\link[=fn_gpsmart]{fn_gpsmart()}}.
\item \code{\link[=fn_gpsmart]{fn_gpsmart()}} filters and ranks \code{input_suspects} based on their probability of committing the \code{input_crime}.
}
}
\author{
Sophie Curtis-Ham
}
