% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prepare_suspect_data.R
\name{fn_prepare_suspect_data}
\alias{fn_prepare_suspect_data}
\title{Prepare suspect data for GP-SMART}
\usage{
fn_prepare_suspect_data(input_suspects_raw)
}
\arguments{
\item{input_suspects_raw}{A data frame with at least one row and 14 columns:
\describe{
\item{\code{person_id}}{A unique reference number for the suspect. Can be a character or numeric vector.}
\item{\code{death_date}}{Suspect's date of death if deceased. A date in format "YYYY-MM-DD"}
\item{\code{x}}{Easting coordinate. Must be in metres to enable distance calculations.}
\item{\code{y}}{Northing coordinate. Must be in metres to enable distance calculations.}
\item{\code{node_category}}{The general category of the activity location. A character vector with values "span" or "event". Span and event nodes are treated differently when calculating activity node attributes.}
\item{\code{node_type}}{The type of activity location. A factor with levels "home", "family_immediate", "family_ip", "family_other", "school", "work", "offence", "victim_witness", "incident", "police_other". Node type determines some activity node attribute values.}
\item{\code{prior_offence_type}}{The type of offence, if it's an offence activity location and one of the crime types for which GP-SMART is currently calibrated. A factor with levels "burglary", "robbery" or "sex", being the crimes GP-SMART is calibrated for. Used for calculating behaviour similarity.}
\item{\code{span_start_date}}{Start date of the activity location, if a span node. A date in format "YYYY-MM-DD".}
\item{\code{span_end_date}}{End date of the activity location, if an event node. A date in format "YYYY-MM-DD".}
\item{\code{event_start_date}}{Start date of the event, if an event node. A date in format "YYYY-MM-DD".}
\item{\code{event_end_date}}{End date of the event, if an event node. A date in format "YYYY-MM-DD".}
\item{\code{event_start_time}}{Start time of the event, if an event node. A difftime in format "HH:MM:SS". Time is not present for "police_other" event nodes in the package data.}
\item{\code{event_end_time}}{End time of the event, if an event node. A difftime in format "HH:MM:SS". Time is not present for "police_other" event nodes in the package data.}
\item{\code{location_type}}{The type of location in which the crime was committed. A factor with levels "residential", "commercial", "public", "street" or "unknown".}
}}
}
\value{
A data frame the same as the input with 7 additional columns:
\itemize{
\item{\code{event_date_time_random}}{Random date-time generated for event nodes.}
\item{\code{daypart}}{The day part of the node. A factor with levels "daytime", "evening", "night" or "all".}
\item{\code{weekpart}}{The week part of the activity node. A factor with levels "weekday", "weekend" or "both".}
\item{\code{spring}}{Whether the node dates include any days in spring (1 = yes, 0 = no)}
\item{\code{summer}}{Whether the node dates include any days in summer (1 = yes, 0 = no)}
\item{\code{autumn}}{Whether the node dates include any days in autumn (1 = yes, 0 = no)}
\item{\code{winter}}{Whether the node dates include any days in winter (1 = yes, 0 = no)}
}
}
\description{
Checks that the minimum necessary variables are present and calculates additional attributes needed for \code{\link[=fn_gpsmart]{fn_gpsmart()}}.
}
\details{
\itemize{
\item {The function will return an error if the input does not contain the correct columns in the correct format.}
\item {Season variables are based on Southern Hemisphere.}
}
}
\examples{
\dontrun{
data(example_input_suspects_raw)
fn_prepare_supect_data(example_input_suspects_raw)
}

}
\seealso{
\itemize{
\item \code{\link[=fn_prepare_input_crime]{fn_prepare_input_crime()}} checks that the minimum necessary variables are present and creates the \code{input_crime} data frame for use in \code{\link[=fn_gpsmart]{fn_gpsmart()}}.
\item \code{\link[=fn_gpsmart]{fn_gpsmart()}} filters and ranks \code{input_suspects} based on their probability of committing the \code{input_crime}.
}
}
\author{
Sophie Curtis-Ham
}
