% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsmartr.R
\docType{package}
\name{gpsmartr-package}
\alias{gpsmartr}
\alias{gpsmartr-package}
\title{gpsmartr: Geographic Profiling - Suspect Mapping And Ranking Technique (GP-SMART) in R}
\description{
Maps input suspect activity locations (nodes) and ranks suspects for an input crime.
}
\details{
gpsmartr.private is a collection of functions to support geographic profiling in crime investigations.
It operationalises GP-SMART (Geographic Profiling: Suspect Mapping and Ranking Technique),
the method described in Curtis-Ham et al (under review) to map and rank suspects for an input crime,
based on the location and attributes of the suspects' activity locations (nodes).
The functions require the user to have a) an input crime and b) a dataset of suspect activity nodes.
These files need to include specific variables describing the location, time and other attributes of both the input crime and suspect activity nodes.
Examples of both files are provided as package data.
The package includes the following functions, to be run in order:
\itemize{
\item \code{\link[=fn_prepare_input_crime]{fn_prepare_input_crime()}} checks that the minimum necessary variables are present in the user's input crime file and creates the \code{input_crime} data frame for use in \code{\link[=fn_gpsmart]{fn_gpsmart()}}.
\item \code{\link[=fn_prepare_suspect_data]{fn_prepare_suspect_data()}} checks that the minimum necessary variables are present in the user's suspect activity nodes file and creates the \code{input_suspects} data frame for use in \code{\link[=fn_gpsmart]{fn_gpsmart()}}.
\item \code{\link[=fn_gpsmart]{fn_gpsmart()}} filters and ranks \code{input_suspects} based on their probability of committing the \code{input_crime}.
\item \code{\link[=fn_map_gpsmart_output]{fn_map_gpsmart_output()}} creates an interactive map visualising the output of \code{\link[=fn_gpsmart]{fn_gpsmart()}}.
}
}
\references{
Curtis-Ham S., Bernasco, W., Medvedev, O. N., & Polaschek, D. L. L (under review).
'A new geographic profiling method for mapping and ranking suspects in crime investigations: GP-SMART'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sophie-c-h/gpsmartr}
  \item Report bugs at \url{https://github.com/Sophie-c-h/gpsmartr/issues}
}

}
\author{
\strong{Maintainer}: Sophie Curtis-Ham \email{sc398@students.waikato.ac.nz} (\href{https://orcid.org/0000-0001-8093-4804}{ORCID})

}
